Set default value for `_serverTextController` in `initState`
Right now the lastKnownConnection value is getting set as the value for the server IP address input every time the widget rebuilds.

The widget will rebuild when you hit `done` on the iOS keyboard.

Setting the value of that field to the `lastKnownConnection` value only once during initalization fixes this.