Rename docker container Accidentally included repo name, shortened to ghcr.io/tarmslitaren/x-haven-server. Rename main-dev tag to just dev. Updated docs
Notes!
This won't create a latest tag until a release is cut, or you could manually upload the latest image.
You will need to manually delete the container with the old name and if you prefer a different container name, now is the time to change it.
Also, if you want to upload to Docker Hub, you would need to set that up with secrets, etc.